requirements.txt
flask>=3.0
openai>=1.0
pyyaml>=6.0
